Change precision parameter type to int in float_type
Modified the `float_type` method to accept an `int` for precision instead of `size_t`. This change ensures consistency across different parts of the codebase and avoids potential issues with type mismatches.

+++ platform/include/roughpy/generics/type.h
@@ -320,7 +320,7 @@ public:
     TypePtr rational_type;
 
     RPY_NO_DISCARD
-    TypePtr float_type(size_t n_precision) const;
+    TypePtr float_type(int n_precision) const;
 
     RPY_NO_DISCARD
     static const MultiPrecisionTypes& get() noexcept;
